# apps/linkmate-backend/internal/users/dto.go
package users

import "time"

type WhoAmIDto struct {
	Id         int       `json:"id"`
	Name       string    `json:"name"`
	Email      string    `json:"email"`
	CreateTime time.Time `json:"createTime"`
	UpdateTime time.Time `json:"updateTime"`
}